chore: remove obsolete stored message interface
