+++ src/application/interfaces/chat/ChatMemory.interface.ts
@@ -1,5 +1,5 @@
-import type { ChatMessage } from '../ai/AIService.interface';
-import type { StoredMessage } from '../messages/StoredMessage.interface';
+import type { ChatMessage } from '../../../domain/messages/ChatMessage.interface';
+import type { StoredMessage } from '../../../domain/messages/StoredMessage.interface';
 
 export interface ChatMemory {
   addMessage(message: StoredMessage): Promise<void>;
