+++ src/application/interfaces/messages/MessageService.interface.ts
@@ -1,7 +1,7 @@
 import type { ServiceIdentifier } from 'inversify';
 
-import type { ChatMessage } from '../ai/AIService.interface';
-import type { StoredMessage } from './StoredMessage.interface';
+import type { ChatMessage } from '../../../domain/messages/ChatMessage.interface';
+import type { StoredMessage } from '../../../domain/messages/StoredMessage.interface';
 
 export interface MessageService {
   addMessage(message: StoredMessage): Promise<void>;
