+++ src/application/use-cases/messages/MessageFactory.ts
@@ -2,8 +2,8 @@ import assert from 'node:assert';
 
 import type { Context } from 'telegraf';
 
+import type { StoredMessage } from '../../../domain/messages/StoredMessage.interface';
 import type { MessageContext } from '../../interfaces/messages/MessageContextExtractor.interface';
-import type { StoredMessage } from '../../interfaces/messages/StoredMessage.interface';
 
 export class MessageFactory {
   static fromUser(ctx: Context, meta: MessageContext): StoredMessage {
