+++ src/domain/messages/ChatMessage.interface.ts
@@ -1,0 +1,13 @@
+export interface ChatMessage {
+  role: 'user' | 'assistant';
+  content: string;
+  username?: string;
+  fullName?: string;
+  replyText?: string;
+  replyUsername?: string;
+  quoteText?: string;
+  userId?: number;
+  messageId?: number;
+  chatId?: number;
+  attitude?: string | null;
+}
