+++ src/infrastructure/persistence/sqlite/SQLiteMessageRepository.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'inversify';
 
-import type { ChatMessage } from '../../../application/interfaces/ai/AIService.interface';
-import type { StoredMessage } from '../../../application/interfaces/messages/StoredMessage.interface';
+import type { ChatMessage } from '../../../domain/messages/ChatMessage.interface';
+import type { StoredMessage } from '../../../domain/messages/StoredMessage.interface';
 import {
   DB_PROVIDER_ID,
   type DbProvider,
