+++ test/RepositoryMessageService.test.ts
@@ -5,7 +5,7 @@ import { type ChatUserRepository } from '../src/domain/repositories/ChatUserRepo
 import { type MessageRepository } from '../src/domain/repositories/MessageRepository.interface';
 import { type UserRepository } from '../src/domain/repositories/UserRepository.interface';
 import { RepositoryMessageService } from '../src/application/use-cases/messages/RepositoryMessageService';
-import { type StoredMessage } from '../src/application/interfaces/messages/StoredMessage.interface';
+import { type StoredMessage } from '../src/domain/messages/StoredMessage.interface';
 import type { LoggerFactory } from '../src/application/interfaces/logging/LoggerFactory.interface';
 
 describe('RepositoryMessageService', () => {
